requirements.txt
requests>=2.32.3
websockets>=14.1
pyopenssl>=25.0.0
pillow>=11.1.0
